requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
cycler==0.11.0
DateTime==4.4
Django==4.1.3
django-binary-database-files==1.0.17
django-cors-headers==3.14.0
djangorestframework==3.14.0
fonttools==4.37.1
google-api-core==2.8.1
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.2
greenlet==2.0.1
gspread==5.4.0
httplib2==0.20.4
idna==3.3
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
protobuf==3.20.1
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygsheets==2.0.5
pyparsing==3.0.9
pyTelegramBotAPI==4.5.1
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
pyzbar==0.1.9
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
threadpoolctl==3.1.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.9
XlsxWriter==3.0.3
xlwings==0.27.10
zope.interface==5.4.0
